+++ handler/healh_check_handler.go
@@ -10,11 +10,12 @@ import (
 )
 
 func RegisterHealthCheckRoutes(e *echo.Echo) {
-	healthCheckGroup := e.Group("/health_check")
-	healthCheckGroup.GET("/ping", healthCheckHandler)
 	e.GET("/", func(c echo.Context) error {
 		return c.String(http.StatusOK, "Welcome to arvigo-backend")
 	})
+
+	healthCheckGroup := e.Group("/health_check")
+	healthCheckGroup.GET("/ping", healthCheckHandler)
 }
 
 func healthCheckHandler(c echo.Context) error {